% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armadillo_forecaster.R
\name{armadillo_forecaster_raw}
\alias{armadillo_forecaster_raw}
\title{Raw forecaster}
\usage{
armadillo_forecaster_raw(
  df,
  forecast_date,
  signals,
  incidence_period = c("epiweek"),
  ahead,
  geo_type = c("state"),
  before_pan,
  mob_shift,
  mob_fun = c("mean", "min", "max"),
  DC,
  initial_fun,
  initial_val,
  lower,
  upper,
  ...
)
}
\arguments{
\item{df}{dataframe returned by covidcast.}

\item{forecast_date}{forecast date, could be a string like "2020-07-01" or parsed date-times in lubridate.}

\item{signals}{signal specifying response variable and safegraph mobility variables.}

\item{incidence_period}{this forecaster only works with epiweek now.}

\item{ahead}{how many epiweeks ahead to predict.}

\item{geo_type}{this forecaster only works with state geo level.}

\item{before_pan}{before_pan if TRUE, the parameter out in Mean.fun is 0, and the first estimated response must be 0.}

\item{mob_shift}{mob_shift the number of epiweek that mobility variable is shifted backwards.}

\item{mob_fun}{mob_fun the function applied to mobility variables, min, mean, or max.}

\item{DC}{DC death delay curve, numerical vector. If NULL, a gamma distribution with scale 3.64, shape 6.28 is used.}

\item{initial_fun}{initial_fun function to initialize the normalized parameter At. It is a function of the sum of death incid}

\item{initial_val}{initial_val initialized values for other model parameters, beta, alpha, mu, sigma.}

\item{lower}{lower lower bound for parameters (At, beta, alpha, mu, sigma).}

\item{upper}{upper upper bound for parameters (At, beta, alpha, mu, sigma).}

\item{...}{... control arguments in optim_sa, such as initial temperature, temperature reduction in outer loop, ...}
}
\value{
point forecasts
}
\description{
Raw forecaster
}
