% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armadillo_forecaster_optimizer.R
\name{fit_optim}
\alias{fit_optim}
\title{Optimizer using simulated annealing}
\usage{
fit_optim(param, DC, y, M, L, before_pan, lower, upper, ...)
}
\arguments{
\item{param}{initial value of arguments in mobility_model (A, beta, alpha, mu, sigma).}

\item{DC}{death delay curve.}

\item{y}{observed response series.}

\item{M}{mobility series.}

\item{L}{length of observed response series.}

\item{before_pan}{same as argument before_pan in mobility_model.}

\item{lower}{a vector of lower bounds (A, beta, alpha, mu, sigma).}

\item{upper}{a vector of upper bounds (A, beta, alpha, mu, sigma).}

\item{...}{control arguments in optim_sa, such as initial temperature, temperature reduction in outer loop, ...}
}
\value{
Best parameters found by simulated annealing.
}
\description{
Optimizer using simulated annealing
}
