% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armadillo_forecaster_optimizer.R
\name{hellinger_loss}
\alias{hellinger_loss}
\title{Hellinger loss function}
\usage{
hellinger_loss(y, yhat)
}
\arguments{
\item{y}{observed response}

\item{yhat}{fitted response}
}
\value{
Hellinger loss
}
\description{
Hellinger loss function
}
