% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armadillo_forecaster_optimizer.R
\name{Mean.fun}
\alias{Mean.fun}
\title{Fit Larry and Valerie's mobility model}
\usage{
Mean.fun(L, At, beta, alpha, mu, sigma, M, DC, before_pan)
}
\arguments{
\item{L}{length of observed response series.}

\item{At}{normalized parameter.}

\item{beta}{mobility function parameter.}

\item{alpha}{mobility function parameter.}

\item{mu}{mobility function parameter.}

\item{sigma}{mobility function parameter.}

\item{M}{mobility series.}

\item{DC}{death delay curve.}

\item{before_pan}{if TRUE, the first estimated response must be 0.}
}
\value{
estmiated response series.
}
\description{
Fit Larry and Valerie's mobility model
}
