% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armadillo_forecaster_helper.R
\name{mob_trans}
\alias{mob_trans}
\title{Transform mobility variables from day to week}
\usage{
mob_trans(x, incidence_length = 7, mob_fun)
}
\arguments{
\item{x}{vector of mobility variables in day.}

\item{incidence_length}{by default, 7. Data in day is transformed to week.}

\item{mob_fun}{function applied to mobility variables. Now this package only supports "min", "max" and "mean".}
}
\value{
Vector of a function of mobility variables in incidence_length.
}
\description{
If the length of the vector is not a multiple of incidence_length, the first a few data points are excluded to make the length of the left a multiple.
}
