% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armadillo_forecaster_optimizer.R
\name{mobility_model}
\alias{mobility_model}
\title{Fit Larry and Valerie's mobility model}
\usage{
mobility_model(L, A, beta, alpha, mu, sigma, M, DC, before_pan)
}
\arguments{
\item{L}{length of observed mobility series.}

\item{A}{normalized parameter. See in the details.}

\item{beta}{mobility function parameter. See in the details.}

\item{alpha}{mobility function parameter. See in the details.}

\item{mu}{mobility function parameter. See in the details.}

\item{sigma}{mobility function parameter. See in the details.}

\item{M}{mobility series. See in the details.}

\item{DC}{death delay curve.}

\item{before_pan}{if TRUE, the first estimated response must be 0. See in the details.}
}
\value{
Estimated response series.
}
\description{
This function fits Larry and Valerie's mobility model. Given a mobility variable series, the estimated death incidence series is returned.
}
\details{
For a single state, let \eqn{Y_t} be the number of deaths at time \eqn{t}. The model can be described as
\deqn{E[Y_t|M = m] = A\sum_{s=1}^{t}exp{s\beta + \alpha G_s}f(t-s)}
for \eqn{t > 1}, where \eqn{M} denotes mobility, \eqn{m = m_1,...,m_t} is the mobility in the past. \eqn{f(t-s)} is the death delay curve, i.e. the probability of someone dying at time \eqn{t} if infected at time \eqn{s}. The probit function is used for \eqn{G_s}.
\deqn{G_s = \sum_{u=1}^{s-1}\Phi((m_u - \mu)/\sigma)}
If before_pan is TRUE, \eqn{E[Y_1|M = m] = 0}, otherwise \eqn{E[Y_1|M = m] = A}.
}
