% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armadillo_forecaster_helper.R
\name{resp_trans}
\alias{resp_trans}
\title{Transform death incidence numbers from day to sum over a week}
\usage{
resp_trans(x, incidence_length = 7)
}
\arguments{
\item{x}{vector of death incidence numbers in day.}

\item{incidence_length}{by default, 7.}
}
\value{
vector of sum of death incidence numbers over the incidence length.
}
\description{
A vector of death incidence numbers in day is transformed to sum over the incidence length. By default, sum over a week.
If the length of the vector is not a multiple of incidence_length, the first a few data points are excluded to make the length of the left a multiple.
}
